requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
packaging==21.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6